fix(stars): refund monitoring and deduplicate story delivery
