--- .github/workflows/pr310-latest-review-fixes.yml
@@ -1,61 +1,0 @@
-name: Validate latest PR 310 review fixes
-
-on:
-  pull_request:
-    branches:
-      - main
-    types:
-      - opened
-      - reopened
-      - synchronize
-
-permissions:
-  contents: write
-
-concurrency:
-  group: pr310-latest-review-fixes-v1
-  cancel-in-progress: true
-
-jobs:
-  patch-test-commit:
-    if: github.event.pull_request.number == 310 && github.actor != 'github-actions[bot]'
-    runs-on: ubuntu-latest
-    timeout-minutes: 25
-    steps:
-      - uses: actions/checkout@v4
-        with:
-          ref: hotfix/stars-payment-safety
-          fetch-depth: 0
-
-      - uses: actions/setup-node@v4
-        with:
-          node-version: 20
-          cache: npm
-
-      - name: Apply latest review fixes
-        run: python3 scripts/pr310-latest-review-fixes.py
-
-      - name: Prepare runtime data directory
-        run: |
-          sudo mkdir -p /data
-          sudo chmod 777 /data
-
-      - name: Install dependencies
-        run: npm ci --legacy-peer-deps
-
-      - name: Build TypeScript
-        run: npm run build
-
-      - name: Run full Jest suite
-        run: timeout 360s npm test -- --runInBand --forceExit
-
-      - name: Commit validated fixes
-        shell: bash
-        run: |
-          rm -f .github/workflows/pr310-latest-review-fixes.yml
-          rm -f scripts/pr310-latest-review-fixes.py
-          git config user.name "chatgpt-codex-connector[bot]"
-          git config user.email "199175422+chatgpt-codex-connector[bot]@users.noreply.github.com"
-          git add -A
-          git commit -m "fix(stars): refund monitoring and deduplicate story delivery"
-          git push origin HEAD:hotfix/stars-payment-safety
